Move all module tests to tests directory
