+++ tests/conftest.py
@@ -1,6 +1,9 @@
 import pytest
 
+from django.conf import settings
 from rest_framework import test
+
+from cloud_ilastik.users.test.factories import UserFactory
 
 
 @pytest.fixture
@@ -11,3 +14,13 @@ def api_rf() -> test.APIRequestFactory:
 @pytest.fixture
 def api_client() -> test.APIClient:
     return test.APIClient()
+
+
+@pytest.fixture(autouse=True)
+def media_storage(settings, tmpdir):
+    settings.MEDIA_ROOT = tmpdir.strpath
+
+
+@pytest.fixture
+def user() -> settings.AUTH_USER_MODEL:
+    return UserFactory()
